pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "can-stealth-attack-ai"
version = "0.1.0"
description = "Laboratoire d'attaque furtive IA sur réseau CAN automobile simulé"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "DanielMBouyou"},
]

dependencies = [
    # CAN
    "python-can>=4.3",
    # Data
    "numpy>=1.26",
    "pandas>=2.1",
    "pyarrow>=14.0",
    # IA
    "torch>=2.1",
    "scikit-learn>=1.3",
    "optuna>=3.4",
    # API
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "websockets>=12.0",
    # Dashboard
    "streamlit>=1.28",
    "plotly>=5.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.7",
    "httpx>=0.25",
]
advanced = [
    "torch-geometric>=2.4",
    "scipy>=1.11",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
